requirements.txt
feedparser>=6.0.11
google-api-python-client>=2.167.0
goose3>=3.1.19
python-dotenv>=1.0.1
PyYAML>=6.0.2
requests>=2.32.3
youtube-transcript-api>=1.0.3
pytest>=8.3.5
